requirements.txt
appdirs==1.4.3
cycler==0.10.0
functools32==3.2.3.post2
matplotlib==2.0.0
mechanize==0.3.1
multi-mechanize==1.2.0
numpy==1.12.1
packaging==16.8
paho-mqtt==1.2
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0